compat+fix: add LPWAVEFORMAT; fix pointer-to-int in huffencode.cpp
- mmsystem.h: add WAVEFORMAT/LPWAVEFORMAT/LPCWAVEFORMAT structs (older
  pre-WAVEFORMATEX audio format used by WWAudio.h)
- huffencode.cpp: replace (int)ptr casts with (ptrdiff_t)(ptr-ptr) for
  correct pointer arithmetic on 64-bit Clang/ARM64

+++ GeneralsMD/Code/Libraries/Source/Compression/EAC/huffencode.cpp
@@ -1050,8 +1050,8 @@ static void HUFF_pack(struct HuffEncodeContext *EC,
 		if (!i3)
 			HUFF_writecode(EC,dest,i);
 
-		if (((int) bptr1- (int) EC->buffer) >= (int)(EC->plen+curpc))
-			curpc = (int) bptr1 - (int) EC->buffer - EC->plen;
+		if ((ptrdiff_t)(bptr1 - EC->buffer) >= (ptrdiff_t)(EC->plen+curpc))
+			curpc = (unsigned int)(bptr1 - EC->buffer) - EC->plen;
 	}
 
 	/* write EOF ([clue] 0gn [10]) */
